pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mr-regression-tools"
version = "0.1.0"
description = "MR Regression Test: English Excel → patch.yml → Japanese Excel updater"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
